refactor(safety-authority): ADR-0035 Stage 3b — move attestation crypto into kirra-safety-authority
Relocates the node-attestation proof-verification module (src/attestation.rs,
issue #73 — INVARIANT #3's per-node Ed25519 challenge-response + the PCR16
measured-boot binding) VERBATIM into the kirra-safety-authority crate. Attestation
is part of the safety-authority boundary the Stage-3 target enumerates, and the
module is pure crypto (no AppState / VerifierStore coupling), so the move touches
no state — it is the next clean slice after 3a's decision surface.

- crates/kirra-safety-authority/src/attestation.rs: the module, moved as-is. The
  single cross-ref `crate::audit_chain::verifying_key_id` is repointed to
  `kirra_audit_hash::verifying_key_id` (the leaf crate where it already lives, slice
  2); `parse_ed25519_public_pem` is re-exported from kirra-audit-hash as before.
- src/attestation.rs is now a `pub use kirra_safety_authority::attestation::*;` shim,
  so every `crate::attestation::…` path (bins, tpm_quote, fleet/operator handlers,
  the ota-installer enroll path, tests) resolves unchanged.
- new crate deps: ed25519-dalek (pkcs8), hex, kirra-audit-hash (prod); base64, hmac,
  sha2, proptest (the attestation test suite's dev-deps). No new package enters the
  tree — the detached-workspace lockfiles (fuzz/, kirra-verifier-pg/, parko/) only
  gain the new dependency edges (refreshed here, --locked lesson from 3a).
- TRACEABILITY_MATRIX.md regenerated: the SG9 attestation tags now resolve at
  crates/kirra-safety-authority/src/attestation.rs (the regen also picks up
  pre-existing bin-handler line drift — the gate scans live source, not the .md).

Behaviour-preserving: verbatim move; the attestation test suite (valid/wrong-key/
tampered/pcr16/legacy-HMAC-rejection/property tests) moved WITH the module and is
the proof — 21/21 in the crate, traceability gate 9/9, full root regression green.
