+++ crates/kirra-safety-authority/src/attestation.rs
@@ -1,0 +1,844 @@
+// src/attestation.rs
+//
+// Node attestation proof verification (issue #73).
+//
+// WHAT CHANGED AND WHY
+// --------------------
+// The prior `verify_attestation` accepted `HMAC-SHA256(KIRRA_ADMIN_TOKEN,
+// nonce)` as the attestation "proof". That key is the SHARED admin token, not
+// a per-node secret — so anyone holding the admin token could attest ANY
+// `node_id` as Trusted. That is admin-ASSERTED trust, not node-PROVEN
+// identity, and the per-node `ak_public_pem` was never read (issue #73, P0).
+//
+// This module replaces that with a per-node challenge-response: the node must
+// prove possession of the PRIVATE attestation key (AK) corresponding to the
+// `ak_public_pem` it registered, by signing the challenge payload for
+// (node_id, nonce) with Ed25519. The verifier checks that signature against
+// the registered public key. Possession of the admin token no longer attests
+// anything — the challenge-response now genuinely "provides its own guarantee".
+//
+// CRYPTO DISCIPLINE (issue #73 prompt):
+//   - No hand-rolled primitives. Signature verification uses the vetted
+//     `ed25519-dalek` crate (already the repo's attestation/federation crypto)
+//     via `VerifyingKey::verify_strict`. SPKI/DER parsing uses the vetted
+//     `pkcs8`/`spki` parser (`from_public_key_der`); only the PEM *armor*
+//     stripping (removing the `-----BEGIN/END-----` lines) is local text
+//     handling, not cryptography.
+//   - No secrets in source/logs/errors. `AttestationError` carries only a
+//     category, never key or signature bytes. Private keys live only in tests,
+//     generated ephemerally at runtime — never committed.
+//   - Fail-closed: a node with no registered AK, a malformed key, a malformed
+//     proof, or a bad signature is REJECTED. There is no accept-by-default.
+//
+// SCOPE / FOLLOW-UP (honest limits — needs human security review):
+//   - The base proof binds (node_id, nonce) via an Ed25519 signature. The
+//     measured-boot follow-up (`verify_attestation_proof_with_pcr16`) now BINDS
+//     the node's presented PCR16 digest INTO that signature and ENFORCES it
+//     against the node's registered `expected_pcr16_digest_hex` (fail-closed:
+//     expected-but-absent or mismatched → `Pcr16Mismatch`). So
+//     `expected_pcr16_digest_hex` is no longer stored-but-unverified.
+//   - LIMIT of THIS path: it authenticates the node's SELF-REPORTED PCR16 under
+//     its AK; it is NOT a hardware TPM *quote* (TPMS_ATTEST, where the TPM itself
+//     signs the live PCR bank), so a node in control of its AK could still sign a
+//     false digest. That deeper root is now CLOSED by `crate::tpm_quote`
+//     (`verify_tpm_quote`), enforced on `/attestation/verify` for a node whose
+//     `node_attestation_policy.require_tpm_quote` is set — see that module.
+//   - The verifier-side change is here; the NODE side must sign the payload
+//     with its AK private key, and real AK/PCR provisioning (TPM / secure
+//     element / KMS) is a DEPLOYMENT concern, not a source change.
+//   - This closes the specific "admin token forges attestation" gap and the
+//     valid/invalid-signature tests pass.
+//
+// NONCE LIFECYCLE (issue #147 — security-review gate, now closed):
+//   The challenge nonce that this proof binds is governed by five invariants.
+//   Together they close the REPLAY and nonce-PREDICTABILITY concerns the #73
+//   note deferred; the crypto half (this module) plus the store/handler half
+//   form the complete challenge-response guarantee:
+//     1. UNPREDICTABLE  — generated by a CSPRNG (`verifier::generate_challenge_nonce`,
+//        OS entropy via `getrandom`), never the wall clock. A time-derived nonce
+//        is guessable and can collide within a nanosecond.
+//     2. SINGLE-USE     — `AppState::consume_challenge` atomically REMOVES the
+//        pending entry; a replayed (already-consumed) nonce finds nothing and is
+//        rejected.
+//     3. TTL-BOUNDED    — entries carry `expires_at_ms` (`CHALLENGE_TTL_MS`);
+//        `consume_challenge` rejects an expired nonce, and `issue_challenge`
+//        prunes expired entries so the bounded store stays clean.
+//     4. NODE-BOUND     — the nonce is keyed by `node_id` in the store AND bound
+//        into the signed payload (`attestation_signing_payload`), so a nonce
+//        issued for node A cannot be consumed for, or replayed as, node B.
+//     5. VERIFY-THEN-CONSUME — the handler verifies this Ed25519 proof BEFORE
+//        calling `consume_challenge`, so a failed proof never burns the pending
+//        nonce (no attacker-forced nonce exhaustion).
+//
+//   STILL OUT OF SCOPE (documented, not silently claimed): measured-boot/TPM
+//   quote binding over `expected_pcr16_digest_hex` (separate follow-up, below);
+//   transport-layer protections (TLS) and side-channel/timing resistance of the
+//   underlying crypto, which are deployment/library concerns, not this gate.
+
+use ed25519_dalek::Signature;
+
+/// Domain-separation tag for the attestation challenge payload. Bumping this
+/// (v1 → v2) is how a future payload-format change is made unambiguous.
+const ATTESTATION_DOMAIN: &[u8] = b"kirra-attestation-challenge-v1";
+
+/// Length of a raw Ed25519 signature in bytes.
+const ED25519_SIGNATURE_LEN: usize = 64;
+
+/// Structured, non-leaking attestation failure reasons.
+///
+/// Each maps to a stable token (`as_str`) suitable for an HTTP error body and
+/// the audit log. Deliberately carries NO key material, signature bytes, or
+/// the proof itself — only the failure category.
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub enum AttestationError {
+    /// The node has no registered `ak_public_pem`. Node-proven attestation is
+    /// impossible without a registered AK — fail closed (this is the core
+    /// #73 refusal: trust can no longer be asserted for an un-keyed node).
+    NoRegisteredKey,
+    /// The registered `ak_public_pem` could not be parsed as an Ed25519
+    /// SubjectPublicKeyInfo public key.
+    MalformedRegisteredKey,
+    /// `proof_hex` was not valid hex, or not a 64-byte Ed25519 signature.
+    MalformedProof,
+    /// The signature did not verify against the registered key over the
+    /// challenge payload for (node_id, nonce).
+    SignatureInvalid,
+    /// The node has a registered `expected_pcr16_digest_hex` (measured-boot
+    /// expectation), but the AK-signed proof did NOT carry a matching presented
+    /// digest — either none was presented, or it did not equal the expectation.
+    /// Fail-closed: a node whose measured boot is unattested or differs from its
+    /// registration is refused, even with a valid (node_id, nonce) signature.
+    Pcr16Mismatch,
+}
+
+impl AttestationError {
+    /// Stable, material-free reason token for HTTP/audit surfaces.
+    #[must_use]
+    pub fn as_str(&self) -> &'static str {
+        match self {
+            Self::NoRegisteredKey => "node has no registered attestation key",
+            Self::MalformedRegisteredKey => "registered attestation key is malformed",
+            Self::MalformedProof => "attestation proof malformed",
+            Self::SignatureInvalid => "attestation signature invalid",
+            Self::Pcr16Mismatch => "attestation pcr16 measured-boot mismatch",
+        }
+    }
+}
+
+/// The exact byte payload the node signs with its AK private key and the
+/// verifier reconstructs to check the signature.
+///
+/// Domain-separated and length-prefixed on `node_id` so two different
+/// `(node_id, nonce)` pairs can never collide on the same payload — a
+/// signature issued for one node/nonce cannot be replayed as another's.
+/// Both the node and the verifier MUST construct this identically.
+#[must_use]
+pub fn attestation_signing_payload(node_id: &str, nonce: u64) -> Vec<u8> {
+    let id = node_id.as_bytes();
+    let mut payload = Vec::with_capacity(ATTESTATION_DOMAIN.len() + 8 + id.len() + 8);
+    payload.extend_from_slice(ATTESTATION_DOMAIN);
+    payload.extend_from_slice(&(id.len() as u64).to_le_bytes());
+    payload.extend_from_slice(id);
+    payload.extend_from_slice(&nonce.to_le_bytes());
+    payload
+}
+
+/// Domain-separated tag appended (with a length-prefixed digest) when a node
+/// BINDS a measured-boot PCR16 digest into its attestation proof. Distinct tag so
+/// a base (node_id, nonce) signature can never be confused with a PCR16-bound one.
+const PCR16_BINDING_DOMAIN: &[u8] = b"|kirra-attestation-pcr16-v1|";
+
+/// The signing payload for an attestation proof that optionally BINDS a
+/// measured-boot PCR16 digest.
+///
+/// - `presented_pcr16_digest_hex == None` → **byte-identical** to
+///   [`attestation_signing_payload`] (the v1 payload), so a node that does not
+///   present a digest, and every pre-PCR16 node, verify unchanged.
+/// - `Some(d)` → the v1 payload followed by the domain tag + a length-prefixed
+///   copy of `d`. The node MUST sign over this exact payload, so the presented
+///   digest is **authenticated by the AK** (not merely asserted alongside it).
+#[must_use]
+pub fn attestation_signing_payload_with_pcr16(
+    node_id: &str,
+    nonce: u64,
+    presented_pcr16_digest_hex: Option<&str>,
+) -> Vec<u8> {
+    let mut payload = attestation_signing_payload(node_id, nonce);
+    if let Some(d) = presented_pcr16_digest_hex {
+        let db = d.as_bytes();
+        payload.extend_from_slice(PCR16_BINDING_DOMAIN);
+        payload.extend_from_slice(&(db.len() as u64).to_le_bytes());
+        payload.extend_from_slice(db);
+    }
+    payload
+}
+
+/// Constant-shape equality for two hex digests (ASCII-case-insensitive, trimmed).
+/// The PCR16 expectation is non-secret (an admin-registered measured-boot value),
+/// so plain comparison is appropriate — the security comes from the AK signature
+/// that BINDS the presented digest, not from hiding the expectation.
+fn pcr16_digest_eq(a: &str, b: &str) -> bool {
+    a.trim().eq_ignore_ascii_case(b.trim())
+}
+
+/// Parse an Ed25519 public key from a PEM-encoded SubjectPublicKeyInfo.
+///
+/// Relocated to the pure `kirra-audit-hash` leaf (a key primitive with no store
+/// dependency) so the persistence-layer `KeyRegistry` and these root attestation /
+/// TPM-quote paths share ONE vetted SPKI parser. Re-exported here so every existing
+/// `crate::attestation::parse_ed25519_public_pem` path is unchanged.
+pub use kirra_audit_hash::parse_ed25519_public_pem;
+
+// ---------------------------------------------------------------------------
+// Operator-proven identity (#314 Phase 1) — the attestation pattern applied to
+// HUMANS. A registered operator proves possession of their Ed25519 private key
+// by signing a challenge, exactly as a node proves possession of its AK. REUSES
+// the same PEM parser, the same domain-separated/length-prefixed canonicalization
+// style, and `verify_strict`.
+// ---------------------------------------------------------------------------
+
+/// Domain tag for an operator clearance-grant signature. Distinct from
+/// `ATTESTATION_DOMAIN` so an operator signature can never be cross-replayed as
+/// a node attestation (or vice versa).
+pub const OPERATOR_GRANT_DOMAIN: &[u8] = b"KIRRA-OPERATOR-GRANT-v1";
+
+/// The exact byte payload an operator signs (with their private key) and the
+/// verifier reconstructs, for a clearance grant on `(operator_id, node_id, nonce)`.
+///
+/// Domain-separated and **length-prefixed on every field** (u64 LE length, then
+/// the field bytes) — the same anti-collision discipline as
+/// [`attestation_signing_payload`], extended to three fields. The browser
+/// (WebCrypto) and the server MUST construct this identically. `nonce` is the
+/// challenge nonce as the hex string the server issued (a string, not a u64, so the
+/// in-browser flow never loses precision on a > 2^53 value).
+#[must_use]
+pub fn operator_grant_signing_payload(operator_id: &str, node_id: &str, nonce: &str) -> Vec<u8> {
+    let mut payload = Vec::with_capacity(
+        OPERATOR_GRANT_DOMAIN.len() + 24 + operator_id.len() + node_id.len() + nonce.len(),
+    );
+    payload.extend_from_slice(OPERATOR_GRANT_DOMAIN);
+    for field in [operator_id, node_id, nonce] {
+        let b = field.as_bytes();
+        payload.extend_from_slice(&(b.len() as u64).to_le_bytes());
+        payload.extend_from_slice(b);
+    }
+    payload
+}
+
+/// Domain tag for an operator EMERGENCY-STOP request (#412 / ADR-0013). DISTINCT
+/// from [`OPERATOR_GRANT_DOMAIN`] so a clearance (RELEASE) signature can NEVER be
+/// replayed as a stop request, nor vice versa — the two verbs ride the same
+/// authenticated channel but are cryptographically non-interchangeable.
+pub const OPERATOR_STOP_DOMAIN: &[u8] = b"KIRRA-OPERATOR-ESTOP-v1";
+
+/// The exact byte payload an operator signs for an EMERGENCY-STOP request on
+/// `(operator_id, node_id, nonce)` (#412 / ADR-0013). Same length-prefixed,
+/// domain-separated discipline as [`operator_grant_signing_payload`], but under
+/// [`OPERATOR_STOP_DOMAIN`] — the stop is the clearance verb INVERTED, and the
+/// distinct domain is what makes a clearance signature unusable as a stop (and
+/// vice versa). The browser (WebCrypto) and the server MUST construct this
+/// identically.
+#[must_use]
+pub fn operator_stop_signing_payload(operator_id: &str, node_id: &str, nonce: &str) -> Vec<u8> {
+    let mut payload = Vec::with_capacity(
+        OPERATOR_STOP_DOMAIN.len() + 24 + operator_id.len() + node_id.len() + nonce.len(),
+    );
+    payload.extend_from_slice(OPERATOR_STOP_DOMAIN);
+    for field in [operator_id, node_id, nonce] {
+        let b = field.as_bytes();
+        payload.extend_from_slice(&(b.len() as u64).to_le_bytes());
+        payload.extend_from_slice(b);
+    }
+    payload
+}
+
+/// Domain tag for a node OTA-adoption report (WS-4). Distinct from the node
+/// attestation and operator domains so an adoption signature can never be
+/// cross-replayed as an attestation proof or a clearance/stop, nor vice versa.
+pub const ADOPTION_REPORT_DOMAIN: &[u8] = b"KIRRA-ADOPTION-REPORT-v1";
+
+/// The exact byte payload a node signs (with its attestation key) and the verifier
+/// reconstructs, to make an OTA-adoption report unforgeable: it binds the reporting
+/// `node_id`, the `applied_digest` it claims to run, and the `reported_at_ms`
+/// timestamp. Domain-separated + length-prefixed on the two string fields (u64 LE
+/// length, then bytes), with `reported_at_ms` appended as a fixed u64 LE — the same
+/// anti-collision discipline as [`operator_grant_signing_payload`]. Binding the
+/// timestamp lets the store reject a replayed OLD report (monotonic upsert).
+#[must_use]
+pub fn adoption_report_signing_payload(
+    node_id: &str,
+    applied_digest: &str,
+    reported_at_ms: u64,
+) -> Vec<u8> {
+    let mut payload = Vec::with_capacity(
+        ADOPTION_REPORT_DOMAIN.len() + 16 + node_id.len() + applied_digest.len() + 8,
+    );
+    payload.extend_from_slice(ADOPTION_REPORT_DOMAIN);
+    for field in [node_id, applied_digest] {
+        let b = field.as_bytes();
+        payload.extend_from_slice(&(b.len() as u64).to_le_bytes());
+        payload.extend_from_slice(b);
+    }
+    payload.extend_from_slice(&reported_at_ms.to_le_bytes());
+    payload
+}
+
+/// Verify an Ed25519 signature over `payload` against a PEM-encoded public key.
+/// Fail-closed: a malformed key, a non-64-byte signature, or a bad signature all
+/// return `false`. Uses `verify_strict` (the same path as node attestation).
+#[must_use]
+pub fn verify_ed25519_pem_signature(pem: &str, payload: &[u8], signature_bytes: &[u8]) -> bool {
+    let Some(vk) = parse_ed25519_public_pem(pem) else {
+        return false;
+    };
+    let Ok(sig_array) = <[u8; ED25519_SIGNATURE_LEN]>::try_from(signature_bytes) else {
+        return false;
+    };
+    vk.verify_strict(payload, &Signature::from_bytes(&sig_array))
+        .is_ok()
+}
+
+/// A short, stable fingerprint for a PEM-encoded Ed25519 public key — the
+/// `audit_chain::verifying_key_id` convention reused for operators. `None` if the
+/// PEM is not a valid Ed25519 SPKI key. Recorded with each grant for
+/// non-repudiation (WHICH key signed it).
+#[must_use]
+pub fn operator_key_fingerprint(pem: &str) -> Option<String> {
+    parse_ed25519_public_pem(pem).map(|vk| kirra_audit_hash::verifying_key_id(&vk))
+}
+
+/// Verify a node attestation proof (issue #73).
+///
+/// The node proves possession of the AK private key matching its registered
+/// `ak_public_pem` by signing [`attestation_signing_payload`] for
+/// `(node_id, nonce)`. Returns `Ok(())` only if a registered key is present,
+/// parses, and the proof is a valid Ed25519 signature over that exact payload.
+/// Every failure path is fail-closed (`Err`), never an accept.
+///
+/// `nonce` single-use / freshness is enforced separately by the caller's
+/// challenge store (`consume_challenge`); this function is the cryptographic
+/// half only.
+// SAFETY: SG9 | REQ: attestation-node-proven-identity | TEST: valid_signature_verifies,wrong_key_is_rejected,tampered_nonce_is_rejected,tampered_node_id_is_rejected,absent_registered_key_fails_closed,malformed_registered_key_fails_closed,malformed_proof_fails_closed,legacy_admin_token_hmac_proof_is_rejected
+// (≅ AEGIS SG-006 node identity. #73: per-node Ed25519 proof against the
+//  registered AK replaces the shared-admin-token HMAC; absent/malformed/
+//  invalid → reject. Verification uses ed25519-dalek verify_strict.)
+pub fn verify_attestation_proof(
+    ak_public_pem: Option<&str>,
+    node_id: &str,
+    nonce: u64,
+    proof_hex: &str,
+) -> Result<(), AttestationError> {
+    // Back-compat: no measured-boot expectation, no presented digest → the
+    // signed payload is byte-identical to v1 (see
+    // `attestation_signing_payload_with_pcr16`), so this is the unchanged #73 path.
+    verify_attestation_proof_with_pcr16(ak_public_pem, node_id, nonce, proof_hex, None, None)
+}
+
+/// Verify a node attestation proof WITH optional measured-boot (PCR16) binding
+/// (the #73 follow-up). Extends [`verify_attestation_proof`] with two parameters:
+///
+/// - `expected_pcr16_digest_hex` — the value the node REGISTERED (`RegisteredNode`),
+///   or `None` if the node carries no measured-boot expectation.
+/// - `presented_pcr16_digest_hex` — the digest the node sent on THIS attestation,
+///   or `None`.
+///
+/// Two enforcements on top of the base Ed25519 (node_id, nonce) proof:
+///   1. **Binding** — the presented digest is folded into the signed payload, so
+///      a node that presents a digest must SIGN over it with its AK; the digest is
+///      authenticated, not merely asserted next to the signature.
+///   2. **Expectation (fail-closed)** — when the node has a registered
+///      `expected_pcr16_digest_hex`, the proof MUST carry a presented digest equal
+///      to it, else `Pcr16Mismatch`. A node with no registered expectation is
+///      unaffected (back-compat).
+///
+/// SCOPE (honest): this binds + enforces the node's SELF-REPORTED PCR16 under its
+/// AK signature and the challenge nonce. It is NOT a hardware TPM *quote* (where
+/// the TPM itself signs the live PCR bank); a node in control of its AK could
+/// still sign a false digest. The full TPMS_ATTEST quote remains the deeper
+/// follow-up — but the registered expectation is no longer dead/unverified.
+// SAFETY: SG9 | REQ: attestation-pcr16-measured-boot-binding | TEST: pcr16_match_with_binding_verifies,pcr16_mismatch_is_rejected,pcr16_expected_but_none_presented_is_rejected,pcr16_digest_must_be_signed_not_just_asserted,pcr16_no_expectation_is_backcompat,pcr16_presented_without_expectation_is_bound
+pub fn verify_attestation_proof_with_pcr16(
+    ak_public_pem: Option<&str>,
+    node_id: &str,
+    nonce: u64,
+    proof_hex: &str,
+    expected_pcr16_digest_hex: Option<&str>,
+    presented_pcr16_digest_hex: Option<&str>,
+) -> Result<(), AttestationError> {
+    // Fail closed: no per-node key registered → cannot prove node identity.
+    let pem = ak_public_pem.ok_or(AttestationError::NoRegisteredKey)?;
+    let verifying_key =
+        parse_ed25519_public_pem(pem).ok_or(AttestationError::MalformedRegisteredKey)?;
+
+    // Decode the proof into a 64-byte Ed25519 signature. (The legacy HMAC
+    // proof is 32 bytes → rejected here as MalformedProof, which is exactly
+    // the gap being closed.)
+    let sig_bytes = hex::decode(proof_hex).map_err(|_| AttestationError::MalformedProof)?;
+    let sig_array: [u8; ED25519_SIGNATURE_LEN] = sig_bytes
+        .as_slice()
+        .try_into()
+        .map_err(|_| AttestationError::MalformedProof)?;
+    let signature = Signature::from_bytes(&sig_array);
+
+    // The signature must cover the presented digest (when one is presented), so a
+    // valid proof AUTHENTICATES the digest under the AK — not just the challenge.
+    let payload =
+        attestation_signing_payload_with_pcr16(node_id, nonce, presented_pcr16_digest_hex);
+    verifying_key
+        .verify_strict(&payload, &signature)
+        .map_err(|_| AttestationError::SignatureInvalid)?;
+
+    // Measured-boot expectation: only enforced when the node registered one.
+    if let Some(expected) = expected_pcr16_digest_hex {
+        match presented_pcr16_digest_hex {
+            Some(presented) if pcr16_digest_eq(presented, expected) => {}
+            // Expected-but-absent OR mismatched → fail closed. (The signature
+            // already proved the presented digest is AK-authenticated; here we
+            // enforce it equals the registration.)
+            _ => return Err(AttestationError::Pcr16Mismatch),
+        }
+    }
+    Ok(())
+}
+
+// ---------------------------------------------------------------------------
+// Tests — ephemeral keypairs generated at runtime; NO committed key material.
+// ---------------------------------------------------------------------------
+#[cfg(test)]
+mod tests {
+    use super::*;
+    // `VerifyingKey`/`B64` are used only by the test helpers here (the shared SPKI
+    // parser moved to `kirra_audit_hash`), so scope them to the test module.
+    use base64::{engine::general_purpose::STANDARD as B64, Engine as _};
+    use ed25519_dalek::{Signer, SigningKey, VerifyingKey};
+    use std::sync::atomic::{AtomicU64, Ordering};
+    use std::time::{SystemTime, UNIX_EPOCH};
+
+    /// #412 — the e-stop and clearance payloads MUST differ for identical inputs
+    /// (distinct domain tags), so a clearance signature can never satisfy an e-stop
+    /// verification and vice versa.
+    #[test]
+    fn stop_and_grant_payloads_are_domain_separated() {
+        let g = operator_grant_signing_payload("alice", "robot-01", "deadbeef");
+        let s = operator_stop_signing_payload("alice", "robot-01", "deadbeef");
+        assert_ne!(
+            g, s,
+            "stop and grant payloads must not collide for the same inputs"
+        );
+        assert!(
+            s.starts_with(OPERATOR_STOP_DOMAIN),
+            "stop payload carries its own domain tag"
+        );
+        assert!(
+            g.starts_with(OPERATOR_GRANT_DOMAIN),
+            "grant payload carries its own domain tag"
+        );
+    }
+
+    /// Per-call counter mixed into the seed so two keypairs created in the
+    /// same nanosecond are still distinct.
+    static SEED_COUNTER: AtomicU64 = AtomicU64::new(0);
+
+    /// Fill an ephemeral 32-byte AK seed at runtime via splitmix64 (seeded
+    /// from the clock + a per-call counter), then derive a signing key. The
+    /// private key exists only for the test's duration and is never written
+    /// to disk or source — an ephemeral test-only keypair, not committed key
+    /// material.
+    fn ephemeral_signing_key() -> SigningKey {
+        let nanos = SystemTime::now()
+            .duration_since(UNIX_EPOCH)
+            .map(|d| d.as_nanos() as u64)
+            .unwrap_or(0xDEAD_BEEF);
+        let mut state = nanos
+            ^ SEED_COUNTER
+                .fetch_add(1, Ordering::Relaxed)
+                .wrapping_mul(0x2545_F491_4F6C_DD1D);
+        let mut seed = [0u8; 32];
+        for chunk in seed.chunks_mut(8) {
+            state = state.wrapping_add(0x9E37_79B9_7F4A_7C15);
+            let mut z = state;
+            z = (z ^ (z >> 30)).wrapping_mul(0xBF58_476D_1CE4_E5B9);
+            z = (z ^ (z >> 27)).wrapping_mul(0x94D0_49BB_1331_11EB);
+            z ^= z >> 31;
+            let bytes = z.to_le_bytes();
+            chunk.copy_from_slice(&bytes[..chunk.len()]);
+        }
+        SigningKey::from_bytes(&seed)
+    }
+
+    /// Build a valid Ed25519 SubjectPublicKeyInfo PEM for a verifying key.
+    /// The 12-byte prefix is the canonical RFC 8410 Ed25519 SPKI header; this
+    /// is test-only encoding of a PUBLIC key (no secret).
+    fn public_key_to_pem(vk: &VerifyingKey) -> String {
+        const ED25519_SPKI_PREFIX: [u8; 12] = [
+            0x30, 0x2a, 0x30, 0x05, 0x06, 0x03, 0x2b, 0x65, 0x70, 0x03, 0x21, 0x00,
+        ];
+        let mut der = ED25519_SPKI_PREFIX.to_vec();
+        der.extend_from_slice(vk.as_bytes());
+        let body = B64.encode(&der);
+        format!("-----BEGIN PUBLIC KEY-----\n{body}\n-----END PUBLIC KEY-----\n")
+    }
+
+    fn sign_proof(sk: &SigningKey, node_id: &str, nonce: u64) -> String {
+        let payload = attestation_signing_payload(node_id, nonce);
+        hex::encode(sk.sign(&payload).to_bytes())
+    }
+
+    /// Sign a proof binding an optional presented PCR16 digest (the v1.1 payload).
+    fn sign_proof_with_pcr16(
+        sk: &SigningKey,
+        node_id: &str,
+        nonce: u64,
+        presented: Option<&str>,
+    ) -> String {
+        let payload = attestation_signing_payload_with_pcr16(node_id, nonce, presented);
+        hex::encode(sk.sign(&payload).to_bytes())
+    }
+
+    const PCR16_X: &str = "abababababababababababababababababababababababababababababababab12";
+    const PCR16_Y: &str = "cdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcd34cd";
+
+    /// HAPPY PATH: registered expectation X, node presents X bound in the
+    /// signature → verifies.
+    #[test]
+    fn pcr16_match_with_binding_verifies() {
+        let sk = ephemeral_signing_key();
+        let pem = public_key_to_pem(&sk.verifying_key());
+        let proof = sign_proof_with_pcr16(&sk, "node-a", 7, Some(PCR16_X));
+        assert_eq!(
+            verify_attestation_proof_with_pcr16(
+                Some(&pem),
+                "node-a",
+                7,
+                &proof,
+                Some(PCR16_X),
+                Some(PCR16_X)
+            ),
+            Ok(())
+        );
+    }
+
+    /// MEASURED-BOOT MISMATCH: the node presents (and correctly signs) a digest Y
+    /// that does not equal the registered expectation X → fail closed.
+    #[test]
+    fn pcr16_mismatch_is_rejected() {
+        let sk = ephemeral_signing_key();
+        let pem = public_key_to_pem(&sk.verifying_key());
+        let proof = sign_proof_with_pcr16(&sk, "node-a", 7, Some(PCR16_Y));
+        assert_eq!(
+            verify_attestation_proof_with_pcr16(
+                Some(&pem),
+                "node-a",
+                7,
+                &proof,
+                Some(PCR16_X),
+                Some(PCR16_Y)
+            ),
+            Err(AttestationError::Pcr16Mismatch)
+        );
+    }
+
+    /// EXPECTED-BUT-ABSENT: a node registered with an expectation that presents NO
+    /// digest is refused, even though its base (node_id, nonce) signature is valid.
+    #[test]
+    fn pcr16_expected_but_none_presented_is_rejected() {
+        let sk = ephemeral_signing_key();
+        let pem = public_key_to_pem(&sk.verifying_key());
+        let proof = sign_proof(&sk, "node-a", 7); // valid v1 signature, no digest
+        assert_eq!(
+            verify_attestation_proof_with_pcr16(
+                Some(&pem),
+                "node-a",
+                7,
+                &proof,
+                Some(PCR16_X),
+                None
+            ),
+            Err(AttestationError::Pcr16Mismatch)
+        );
+    }
+
+    /// SECURITY CORE: the presented digest must be SIGNED, not merely asserted.
+    /// Here the node signs only the base payload (no binding) but the request
+    /// asserts the correct digest X. Because the signature does not cover X, it
+    /// fails to verify against the bound payload → SignatureInvalid (not accepted).
+    #[test]
+    fn pcr16_digest_must_be_signed_not_just_asserted() {
+        let sk = ephemeral_signing_key();
+        let pem = public_key_to_pem(&sk.verifying_key());
+        let proof = sign_proof(&sk, "node-a", 7); // signs the BASE payload only
+        assert_eq!(
+            verify_attestation_proof_with_pcr16(
+                Some(&pem),
+                "node-a",
+                7,
+                &proof,
+                Some(PCR16_X),
+                Some(PCR16_X)
+            ),
+            Err(AttestationError::SignatureInvalid)
+        );
+    }
+
+    /// BACK-COMPAT: no expectation + no presented digest → byte-identical to v1,
+    /// so a pre-PCR16 node verifies unchanged (and the 4-arg wrapper matches).
+    #[test]
+    fn pcr16_no_expectation_is_backcompat() {
+        let sk = ephemeral_signing_key();
+        let pem = public_key_to_pem(&sk.verifying_key());
+        let proof = sign_proof(&sk, "node-a", 7);
+        assert_eq!(
+            verify_attestation_proof_with_pcr16(Some(&pem), "node-a", 7, &proof, None, None),
+            Ok(())
+        );
+        // The 4-arg path is exactly this case.
+        assert_eq!(
+            verify_attestation_proof(Some(&pem), "node-a", 7, &proof),
+            Ok(())
+        );
+        // And the bound payload with no digest equals the v1 payload byte-for-byte.
+        assert_eq!(
+            attestation_signing_payload_with_pcr16("node-a", 7, None),
+            attestation_signing_payload("node-a", 7)
+        );
+    }
+
+    /// A node with NO registered expectation MAY still present a digest; it is
+    /// bound into the signature (authenticated) but not compared → verifies.
+    #[test]
+    fn pcr16_presented_without_expectation_is_bound() {
+        let sk = ephemeral_signing_key();
+        let pem = public_key_to_pem(&sk.verifying_key());
+        let proof = sign_proof_with_pcr16(&sk, "node-a", 7, Some(PCR16_X));
+        assert_eq!(
+            verify_attestation_proof_with_pcr16(
+                Some(&pem),
+                "node-a",
+                7,
+                &proof,
+                None,
+                Some(PCR16_X)
+            ),
+            Ok(())
+        );
+    }
+
+    /// CRYPTO CORRECTNESS: a valid signature over the challenge verifies.
+    #[test]
+    fn valid_signature_verifies() {
+        let sk = ephemeral_signing_key();
+        let pem = public_key_to_pem(&sk.verifying_key());
+        let proof = sign_proof(&sk, "node-a", 42);
+        assert_eq!(
+            verify_attestation_proof(Some(&pem), "node-a", 42, &proof),
+            Ok(())
+        );
+    }
+
+    /// LEGITIMATE ENROLLMENT: a second independent node enrolls with its own
+    /// key and succeeds — the happy path is preserved for properly-keyed nodes.
+    #[test]
+    fn independent_node_enrollment_succeeds() {
+        let sk = ephemeral_signing_key();
+        let pem = public_key_to_pem(&sk.verifying_key());
+        let proof = sign_proof(&sk, "lidar-front", 7);
+        assert!(verify_attestation_proof(Some(&pem), "lidar-front", 7, &proof).is_ok());
+    }
+
+    /// CRYPTO CORRECTNESS: a signature from a DIFFERENT key is rejected
+    /// against the registered key.
+    #[test]
+    fn wrong_key_is_rejected() {
+        let registered = ephemeral_signing_key();
+        let attacker = ephemeral_signing_key();
+        let pem = public_key_to_pem(&registered.verifying_key());
+        let forged = sign_proof(&attacker, "node-a", 42);
+        assert_eq!(
+            verify_attestation_proof(Some(&pem), "node-a", 42, &forged),
+            Err(AttestationError::SignatureInvalid)
+        );
+    }
+
+    /// REPLAY/BINDING: a signature for nonce N must not verify for nonce N+1.
+    #[test]
+    fn tampered_nonce_is_rejected() {
+        let sk = ephemeral_signing_key();
+        let pem = public_key_to_pem(&sk.verifying_key());
+        let proof = sign_proof(&sk, "node-a", 100);
+        assert_eq!(
+            verify_attestation_proof(Some(&pem), "node-a", 101, &proof),
+            Err(AttestationError::SignatureInvalid)
+        );
+    }
+
+    /// DOMAIN BINDING: a signature for node "a" must not verify for node "b"
+    /// (payload binds node_id).
+    #[test]
+    fn tampered_node_id_is_rejected() {
+        let sk = ephemeral_signing_key();
+        let pem = public_key_to_pem(&sk.verifying_key());
+        let proof = sign_proof(&sk, "node-a", 42);
+        assert_eq!(
+            verify_attestation_proof(Some(&pem), "node-b", 42, &proof),
+            Err(AttestationError::SignatureInvalid)
+        );
+    }
+
+    /// FAIL-CLOSED: a node with no registered AK cannot attest. This is the
+    /// core #73 refusal — trust can no longer be asserted for an un-keyed node.
+    #[test]
+    fn absent_registered_key_fails_closed() {
+        let sk = ephemeral_signing_key();
+        let proof = sign_proof(&sk, "node-a", 42);
+        assert_eq!(
+            verify_attestation_proof(None, "node-a", 42, &proof),
+            Err(AttestationError::NoRegisteredKey)
+        );
+    }
+
+    /// FAIL-CLOSED: a malformed registered key is rejected, not bypassed.
+    #[test]
+    fn malformed_registered_key_fails_closed() {
+        let sk = ephemeral_signing_key();
+        let proof = sign_proof(&sk, "node-a", 42);
+        assert_eq!(
+            verify_attestation_proof(
+                Some("-----BEGIN PUBLIC KEY-----\nnot-base64!!\n-----END PUBLIC KEY-----"),
+                "node-a",
+                42,
+                &proof
+            ),
+            Err(AttestationError::MalformedRegisteredKey)
+        );
+        // A syntactically-valid-but-wrong-algorithm / truncated DER also fails.
+        assert_eq!(
+            verify_attestation_proof(
+                Some("-----BEGIN PUBLIC KEY-----\nAAAA\n-----END PUBLIC KEY-----"),
+                "node-a",
+                42,
+                &proof
+            ),
+            Err(AttestationError::MalformedRegisteredKey)
+        );
+    }
+
+    /// FAIL-CLOSED: a malformed proof (bad hex / wrong length) is rejected.
+    #[test]
+    fn malformed_proof_fails_closed() {
+        let sk = ephemeral_signing_key();
+        let pem = public_key_to_pem(&sk.verifying_key());
+        assert_eq!(
+            verify_attestation_proof(Some(&pem), "node-a", 42, "zzzz"),
+            Err(AttestationError::MalformedProof)
+        );
+        assert_eq!(
+            verify_attestation_proof(Some(&pem), "node-a", 42, "abcd"),
+            Err(AttestationError::MalformedProof)
+        );
+    }
+
+    /// GAP CLOSED: the pre-#73 proof — `HMAC-SHA256(admin_token, nonce)`,
+    /// which anyone holding the admin token could compute for ANY node — is
+    /// now rejected. (It is a 32-byte MAC, not a 64-byte Ed25519 signature,
+    /// and it is not a signature by the node's AK regardless.)
+    #[test]
+    fn legacy_admin_token_hmac_proof_is_rejected() {
+        use hmac::digest::KeyInit; // hmac 0.13: `new_from_slice` moved to `KeyInit`
+        use hmac::{Hmac, Mac};
+        use sha2::Sha256;
+        type HmacSha256 = Hmac<Sha256>;
+
+        let sk = ephemeral_signing_key();
+        let pem = public_key_to_pem(&sk.verifying_key());
+
+        // Reconstruct exactly what the old verifier accepted.
+        let admin_token = "any-admin-token-value";
+        let nonce: u64 = 42;
+        let mut mac = HmacSha256::new_from_slice(admin_token.as_bytes()).unwrap();
+        mac.update(&nonce.to_le_bytes());
+        let legacy_proof = hex::encode(mac.finalize().into_bytes());
+
+        let result = verify_attestation_proof(Some(&pem), "node-a", nonce, &legacy_proof);
+        assert!(
+            result.is_err(),
+            "the legacy admin-token HMAC proof must NOT be accepted (#73); got {result:?}"
+        );
+        // Specifically rejected at the length gate (HMAC-SHA256 is 32 bytes).
+        assert_eq!(result, Err(AttestationError::MalformedProof));
+    }
+
+    /// The signing payload is deterministic and bound to (node_id, nonce).
+    #[test]
+    fn payload_is_deterministic_and_bound() {
+        assert_eq!(
+            attestation_signing_payload("n", 1),
+            attestation_signing_payload("n", 1)
+        );
+        assert_ne!(
+            attestation_signing_payload("n", 1),
+            attestation_signing_payload("n", 2)
+        );
+        assert_ne!(
+            attestation_signing_payload("a", 1),
+            attestation_signing_payload("b", 1)
+        );
+        // Length-prefixing prevents a node_id/nonce boundary-shift collision.
+        assert_ne!(
+            attestation_signing_payload("ab", 1),
+            attestation_signing_payload("a", 1)
+        );
+    }
+
+    // -----------------------------------------------------------------------
+    // PROPERTY: the signing payload is INJECTIVE over (node_id, nonce).
+    //
+    // INVARIANT: distinct (node_id, nonce) pairs NEVER produce the same signing
+    // payload. SOURCE: the construction is a domain tag, then the node_id
+    // LENGTH (u64 LE), then the node_id bytes, then the nonce (u64 LE). The
+    // length prefix makes the (node_id || nonce) boundary unambiguous, so the
+    // map is injective. Why it matters: if two different (node_id, nonce) pairs
+    // shared a payload, a signature issued for one could be replayed for the
+    // other — a cross-node / cross-nonce attestation forgery. This property is
+    // the cryptographic justification for the length-prefix.
+    // -----------------------------------------------------------------------
+    use proptest::prelude::*;
+
+    proptest! {
+        #![proptest_config(ProptestConfig::with_cases(4000))]
+
+        /// Different inputs ⇒ different payloads (injectivity / no collisions).
+        #[test]
+        fn prop_signing_payload_is_injective(
+            id_a in "[ -~]{0,40}",
+            n_a in any::<u64>(),
+            id_b in "[ -~]{0,40}",
+            n_b in any::<u64>(),
+        ) {
+            let same_input = id_a == id_b && n_a == n_b;
+            let same_payload = attestation_signing_payload(&id_a, n_a)
+                == attestation_signing_payload(&id_b, n_b);
+            // Equal payloads are permitted ONLY when the inputs are equal.
+            prop_assert_eq!(
+                same_payload, same_input,
+                "payload collision across distinct inputs: ({:?},{}) vs ({:?},{})",
+                id_a, n_a, id_b, n_b
+            );
+        }
+
+        /// A specific stress on the boundary-shift class: moving a byte across
+        /// the node_id/nonce seam must change the payload (length-prefix guard).
+        #[test]
+        fn prop_no_boundary_shift_collision(
+            head in "[ -~]{1,20}",
+            tail in "[ -~]{1,20}",
+            nonce in any::<u64>(),
+        ) {
+            let combined = format!("{head}{tail}");
+            // Same concatenated id-bytes split at two different points must
+            // still yield distinct payloads unless they are literally equal.
+            let p1 = attestation_signing_payload(&combined, nonce);
+            let p2 = attestation_signing_payload(&head, nonce);
+            prop_assert_ne!(p1, p2);
+        }
+    }
+}
